+++ apps/web/src/activities/calendar/index.ts
@@ -6,7 +6,7 @@ import ru from "./translations/ru.json"
 export default createActivityExtension("calendar", {
   Component: () => import("./components"),
   Icon: () => import("./components/icon"),
-  routes: ["/calendar", "/calendar/:extension"],
+  routes: ["/calendar", "/calendar/:view"],
   readableName: "@ordo-activity-calendar/title",
   translations: { en, ru },
 })
